Extracted interface
