+++ Src/Kronos.Core/Messages/IRequest.cs
@@ -1,0 +1,8 @@
+﻿using Kronos.Core.Serialization;
+
+namespace Kronos.Core.Messages
+{
+    public interface IRequest : ISerializable<object>
+    {
+    }
+}
